+++ demos/hello_stateless/example/app.jsx
@@ -1,0 +1,7 @@
+/**
+ * Using the Hello component
+ */
+const Hello = require('../index.jsx');
+const markup = <Hello name="Juanma" style={{color: 'red'}}/>;
+
+ReactDOM.render(markup, document.getElementById('container'));