+++ grunt/cleanTask.js
@@ -1,3 +1,12 @@
+const config = require('../config.js');
+
 module.exports = {
-  public: ['example/server/public']
+
+  /**
+   * Drop the distribution folder. This task runs first when all the build
+   * process start.
+   *
+   * Use: `grunt clean:public`
+   */
+  public: [`${config.dist}`]
 };